# routes/user1.js
const express = require("express");
const { check, validationResult } = require("express-validator");
const bcrypt = require("bcryptjs");
const jwt = require("jsonwebtoken");
const router = express.Router();
const auth = require("../middleware/auth");
const sample = require("../middleware/sample");
const { models } = require("../models");
const User = models.User;
const { createProfile, updateProfile, getProfile } = require("../controller/user");

router.post('/profile', auth, createProfile);
router.put('/profile', auth, updateProfile);
router.get('/profile', auth, getProfile);

/**
 * @method - POST
 * @param - /signup
 * @description - User SignUp
 */
 router.post(
    "/signup",
    [
      check("username", "Please Enter a Valid Username")
        .not()
        .isEmpty(),
      check("email", "Please enter a valid email").isEmail(),
      check("password", "Please enter a valid password").isLength({
        min: 6
      })
    ],
    async (req, res) => {
      console.log("POST /user/signup passing initial check");
      const errors = validationResult(req);
      if (!errors.isEmpty()) {
        return res.status(400).json({
          errors: errors.array()
        });
      }
      // console.log("POST /user/signup ", errors);
  
      const { username, email, password } = req.body;
      try {
        let user = await User.findOne({
          where: {email}
        });
        if (user) {
          return res.status(400).json({
            msg: "User Already Exists"
          });
        }
  
        user = new User({
          username,
          email,
          password
        });
  
        const salt = await bcrypt.genSalt(10);
        user.password = await bcrypt.hash(password, salt);
  
        await user.save();
  
        const payload = {
          user: {
            id: user.id
          }
        };
  
        jwt.sign(
          payload,
          process.env.secretKey,
          {
            expiresIn: 10000
          },
          (err, token) => {
            if (err) throw err;
            res.status(201).json({
              token
            });
          }
        );
      } catch (err) {
        console.log(err.message);
        res.status(500).send("Error in Saving");
      }
    }
  );
  
  router.post(
    "/login",
    [
      check("email", "Please enter a valid email").isEmail(),
      check("password", "Please enter a valid password").isLength({
        min: 6
      })
    ],
    async (req, res) => {
      const errors = validationResult(req);
  
      if (!errors.isEmpty()) {
        return res.status(400).json({
          errors: errors.array()
        });
      }
  
      const { email, password } = req.body;
      try {
        let user = await User.findOne({
          where: {email}
        });
        if (!user)
          return res.status(400).json({
            message: "User Not Exist"
          });
  
        const isMatch = await bcrypt.compare(password, user.password);
        if (!isMatch)
          return res.status(400).json({
            message: "Incorrect Password !"
          });
  
        const payload = {
          user: {
            id: user.id
          }
        };

        // console.log("JWT login: ", payload, process.env.secretKey);
  
        jwt.sign(
          payload,
          process.env.secretKey,
          {
            expiresIn: 3600
          },
          (err, token) => {
            if (err) throw err;
            res.status(200).json({
              user, token
            });
          }
        );
      } catch (e) {
        console.error(e);
        res.status(500).json({
          message: "Server Error"
        });
      }
    }
  );
  
  /**
   * @method - GET
   * @description - Get LoggedIn User
   * @param - /user/me
   */
  router.get("/me", auth, async (req, res) => {
    console.log('router to get me route hit')
    try {
      // request.user is getting fetched from Middleware after token authentication
      const user = await User.findById(req.user.id);
      res.status(200).json(user);
    } catch (e) {
      res.send({ message: "Error in Fetching user" });
    }
  });

module.exports = router;